Add Dockerfile to be able to develop in a Container
This Container is supposed to be used with the VSCode Remote Container extension.
It has every required tools ready to be used at the proper location.
It copy inside the container all the .apk files that are in the folder where you start the Remote Container extension.
You can either attach a running container or build it from the Dockerfile.

Signed-off-by: Rémi Lavedrine <remi.lavedrine@outlook.com>